taurus_monitoring/requirements.txt
alembic==0.6.7
enum34==1.0.4
boto==2.38.0
mysql-python==1.2.5
nta.utils
requests==2.0.1
sqlalchemy==0.9.4
